# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PropTech API
  description: API REST para la plataforma PropTech (Alquiler & Venta con Red Social de Reputación)
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Entorno de Desarrollo Local
paths:
...
  /saved-searches:
    get:
      summary: Listar búsquedas guardadas del usuario autenticado
      tags:
        - Alertas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de búsquedas guardadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedSearchDTO'
        '401':
          description: No autorizado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    post:
      summary: Crear una nueva búsqueda guardada
      tags:
        - Alertas
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearchCreateRequest'
      responses:
        '201':
          description: Búsqueda guardada creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchDTO'
        '401':
          description: No autorizado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: Límite de búsquedas guardadas alcanzado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /saved-searches/{id}:
    patch:
      summary: Activar o desactivar una búsqueda guardada
      tags:
        - Alertas
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - active
              properties:
                active:
                  type: boolean
      responses:
        '200':
          description: Búsqueda guardada actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSearchDTO'
        '401':
          description: No autorizado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Acceso denegado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Búsqueda guardada no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      summary: Eliminar una búsqueda guardada
      tags:
        - Alertas
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Búsqueda guardada eliminada
        '401':
          description: No autorizado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Acceso denegado
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Búsqueda guardada no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    GeoJsonGeometry:
      type: object
      description: Objeto GeoJSON de geometría (Polygon o MultiPolygon)
      properties:
        type:
          type: string
          enum: [Polygon, MultiPolygon]
        coordinates:
          type: array
          items: {}
    PropertySearchRequest:
      type: object
      description: Parámetros de búsqueda avanzada de inmuebles
      properties:
        polygon:
          $ref: '#/components/schemas/GeoJsonGeometry'
          description: Polígono GeoJSON opcional. Si presente, usa ST_Intersects; si no, usa ST_DWithin.
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        radius:
          type: number
          format: float
          description: Radio en metros (usado con lat/lng cuando no hay polygon)
        minPrice:
          type: number
        maxPrice:
          type: number
        minRooms:
          type: integer
        page:
          type: integer
          default: 0
        size:
          type: integer
          default: 20
...
    SavedSearchDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        filters:
          $ref: '#/components/schemas/PropertySearchRequest'
        active:
          type: boolean
        lastCheckedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SavedSearchCreateRequest:
      type: object
      required:
        - name
        - filters
      properties:
        name:
          type: string
          maxLength: 100
        filters:
          $ref: '#/components/schemas/PropertySearchRequest'